chore: fmt

+++ test/recon-curvance/TargetFunctions.sol
@@ -7,19 +7,16 @@ import {Properties} from "./Properties.sol";
 import {vm} from "@chimera/Hevm.sol";
 import {IUniV2Pool} from "src/IUniV2Pool.sol";
 
-
 abstract contract TargetFunctions is BaseTargetFunctions, Properties, BeforeAfter {
-
-
     function pool_mint(address to) public {
         IUniV2Pool(uniV2Pool).mint(to);
     }
-    
+
     function burn(address to) public {
         IUniV2Pool(uniV2Pool).burn(to);
     }
 
-    function swap(uint amount0Out, uint amount1Out, address to) public {
+    function swap(uint256 amount0Out, uint256 amount1Out, address to) public {
         IUniV2Pool(uniV2Pool).swap(amount0Out, amount1Out, to, "");
     }
 
@@ -34,12 +31,13 @@ abstract contract TargetFunctions is BaseTargetFunctions, Properties, BeforeAfte
     function donateToken0(uint256 amt) public {
         tokenA.transfer(uniV2Pool, amt);
     }
+
     function donateToken1(uint256 amt) public {
         tokenB.transfer(uniV2Pool, amt);
     }
+
     function donateBoth(uint256 amt) public {
         tokenA.transfer(uniV2Pool, amt);
         tokenB.transfer(uniV2Pool, amt);
     }
-
 }
